updated my serviceimplementation

+++ src/main/java/org/example/ecommerce/Dto/ProductRequestDto.java
@@ -3,15 +3,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductRequestDto {
     private String name;
-    private BigDecimal price;
-    private Integer stockQuantity;
+    private double price;
+    private int stockQuantity;
     private Long categoryId;
 
 }
